requirements.txt
Click==7.0
cycler==0.10.0
Flask==1.1.1
gunicorn==20.0.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.17.4
opencv-python-headless==4.1.2.30
Pillow==6.2.1
pkg-resources==0.0.0
pyparsing==2.4.5
python-dateutil==2.8.1
six==1.13.0
Werkzeug==0.16.0
